Generalise input/output types of a process
The type of the stdin event is generalised from ByteString to a
The type of a stdout event is generalised from ByteString to b

The motiviation for this change is that it's easier to implement
complicated ByteString parsing logic by directly manipulating the handle
rather than trying to manipulate the resulting `Event t ByteString`
which the old interface provided.

The changes also make it possible for the decoding of the output of a
process to depend on the values which have appear in the input. I have
used this when interacting with a language server.